feat: replace user card to finish help button
